Rename error components, update errors pages fiel structure, delete dummy md files
