--- src/app/pages/errors-overview.md
@@ -1,127 +1,0 @@
----
-title: Errors Overview
-slug: errors-overview
-description: Errors Overview for Angular Developers
-meta:
-  - name: title
-    content: Errors Overview Title from meta
-  - property: og:title
-    content: About
-  - property: og:description
-    content: Errors Overview Description from meta
-  - property: og:url
-    content: https://errors.angular.io/NG0100
-  - property: twitter:card
-    content: summary
-  - property: twitter:title
-    content: "errors overview  for twitter"
-  - property: twitter:description
-    content: Error NG0100 Unreachable Code Detected
-  - name: twitter:image
-    content: "/errors-overview.png"
-  - property: og:image
-    content: "/errors-overview.png"
----
-
-<div class="prose">
-
-## API Report File for "angular-srcs"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-// @public (undocumented)
-export enum ErrorCode {
-  COMPONENT_IMPORT_NOT_STANDALONE = 2011,
-  COMPONENT_INVALID_SHADOW_DOM_SELECTOR = 2009,
-  // (undocumented)
-  COMPONENT_MISSING_TEMPLATE = 2001,
-  COMPONENT_NOT_STANDALONE = 2010,
-  COMPONENT_RESOURCE_NOT_FOUND = 2008,
-  COMPONENT_UNKNOWN_IMPORT = 2012,
-  // (undocumented)
-  CONFIG_EXTENDED_DIAGNOSTICS_IMPLIES_STRICT_TEMPLATES = 4003,
-  // (undocumented)
-  CONFIG_EXTENDED_DIAGNOSTICS_UNKNOWN_CATEGORY_LABEL = 4004,
-  // (undocumented)
-  CONFIG_EXTENDED_DIAGNOSTICS_UNKNOWN_CHECK = 4005,
-  // (undocumented)
-  CONFIG_FLAT_MODULE_NO_INDEX = 4001,
-  // (undocumented)
-  CONFIG_STRICT_TEMPLATES_IMPLIES_FULL_TEMPLATE_TYPECHECK = 4002,
-  CONFLICTING_INPUT_TRANSFORM = 2020,
-  // (undocumented)
-  DECORATOR_ARG_NOT_LITERAL = 1001,
-  // (undocumented)
-  DECORATOR_ARITY_WRONG = 1002,
-  DECORATOR_COLLISION = 1006,
-  // (undocumented)
-  DECORATOR_NOT_CALLED = 1003,
-  // (undocumented)
-  DECORATOR_UNEXPECTED = 1005,
-  DIRECTIVE_INHERITS_UNDECORATED_CTOR = 2006,
-  // (undocumented)
-  DIRECTIVE_MISSING_SELECTOR = 2004,
-  DUPLICATE_VARIABLE_DECLARATION = 8006,
-  HOST_BINDING_PARSE_ERROR = 5001,
-  HOST_DIRECTIVE_COMPONENT = 2015,
-  HOST_DIRECTIVE_CONFLICTING_ALIAS = 2018,
-  HOST_DIRECTIVE_INVALID = 2013,
-  HOST_DIRECTIVE_MISSING_REQUIRED_BINDING = 2019,
-  HOST_DIRECTIVE_NOT_STANDALONE = 2014,
-  HOST_DIRECTIVE_UNDEFINED_BINDING = 2017,
-  IMPORT_CYCLE_DETECTED = 3003,
-  IMPORT_GENERATION_FAILURE = 3004,
-  INJECTABLE_DUPLICATE_PROV = 9001,
-  INJECTABLE_INHERITS_INVALID_CONSTRUCTOR = 2016,
-  INLINE_TCB_REQUIRED = 8900,
-  INLINE_TYPE_CTOR_REQUIRED = 8901,
-  INVALID_BANANA_IN_BOX = 8101,
-  LOCAL_COMPILATION_IMPORTED_STYLES_STRING = 11002,
-  LOCAL_COMPILATION_IMPORTED_TEMPLATE_STRING = 11001,
-  MISSING_CONTROL_FLOW_DIRECTIVE = 8103,
-  MISSING_NGFOROF_LET = 8105,
-  MISSING_PIPE = 8004,
-  MISSING_REFERENCE_TARGET = 8003,
-  MISSING_REQUIRED_INPUTS = 8008,
-  NGMODULE_BOOTSTRAP_IS_STANDALONE = 6009,
-  NGMODULE_DECLARATION_IS_STANDALONE = 6008,
-  NGMODULE_DECLARATION_NOT_UNIQUE = 6007,
-  NGMODULE_INVALID_DECLARATION = 6001,
-  NGMODULE_INVALID_EXPORT = 6003,
-  NGMODULE_INVALID_IMPORT = 6002,
-  NGMODULE_INVALID_REEXPORT = 6004,
-  NGMODULE_MODULE_WITH_PROVIDERS_MISSING_GENERIC = 6005,
-  NGMODULE_REEXPORT_NAME_COLLISION = 6006,
-  NGMODULE_VE_DEPENDENCY_ON_IVY_LIB = 6999,
-  NULLISH_COALESCING_NOT_NULLABLE = 8102,
-  OPTIONAL_CHAIN_NOT_NULLABLE = 8107,
-  // (undocumented)
-  PARAM_MISSING_TOKEN = 2003,
-  // (undocumented)
-  PIPE_MISSING_NAME = 2002,
-  SCHEMA_INVALID_ATTRIBUTE = 8002,
-  SCHEMA_INVALID_ELEMENT = 8001,
-  SKIP_HYDRATION_NOT_STATIC = 8108,
-  SPLIT_TWO_WAY_BINDING = 8007,
-  SUFFIX_NOT_SUPPORTED = 8106,
-  SUGGEST_STRICT_TEMPLATES = 10001,
-  SUGGEST_SUBOPTIMAL_TYPE_INFERENCE = 10002,
-  // (undocumented)
-  SYMBOL_NOT_EXPORTED = 3001,
-  TEMPLATE_PARSE_ERROR = 5002,
-  TEXT_ATTRIBUTE_NOT_BINDING = 8104,
-  UNDECORATED_CLASS_USING_ANGULAR_FEATURES = 2007,
-  UNDECORATED_PROVIDER = 2005,
-  // (undocumented)
-  VALUE_HAS_WRONG_TYPE = 1010,
-  // (undocumented)
-  VALUE_NOT_LITERAL = 1011,
-  WARN_NGMODULE_ID_UNNECESSARY = 6100,
-  WRITE_TO_READ_ONLY_VARIABLE = 8005,
-}
-
-// (No @packageDocumentation comment for this package)
-```
-
-</div>
