+++ src/app/pages/errors.page.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+
+@Component({
+  standalone: true,
+  imports: [RouterOutlet],
+  template: ` <router-outlet /> `,
+})
+export default class ErrorsPageComponent {}
