--- src/content/errors/0200.md
@@ -1,108 +1,0 @@
----
-title: "0200"
-slug: "0200"
-description: Error NG0200 Circular Dependency Detected
-author: Nelson Gutierrez
-authorTwitter: nelsongutide
-lastUpdated: Jun 12, 2023
-meta:
-  - name: NG-0200
-    content: NG-0200 Error
-  - property: og:title
-    content: "0200: NG0200 Error Unreachable Code Detected og title"
-  - property: og:description
-    content: Error NG0200 Unreachable Code Detected og description
-  - property: og:image
-    content: "/ng-0200.png"
-  - property: twitter:title
-    content: "0200: NG0200 twitter Error Unreachable Code Detected for twitter"
-  - property: twitter:description
-    content: Error NG0200 twitter Unreachable Code Detected test for
-  - name: twitter:image
-    content: "/ng-0200.png"
-  - name: twitter:card
-    property: summary_large_image
----
-
-# Angular Error NG0200 Demystified: Circular Dependency Detected
-
-As an Angular developer, you're no stranger to the intricate nature of web application development. Angular applications can become quite complex, and with complexity often comes a variety of challenges and errors. In this article, we'll delve into Angular Error NG0200, which pertains to circular dependencies, and we'll explore its significance, causes, and strategies for resolving it.
-
-## Unraveling NG0200 Error
-
-Angular Error NG0200 is a linting error that highlights the presence of circular dependencies within your application. The error message may resemble the following:
-
-```
-Error NG0200: Circular dependency in DI detected.
-```
-
-Though this error message might seem cryptic initially, let's break it down to understand its essence.
-
-## Interpreting the Error Message
-
-1. **`Error NG0200:`** This unique identifier helps Angular developers recognize the specific error at a glance, making it easier to diagnose.
-
-2. **`Circular dependency in DI detected.`** This portion of the error message signifies that a circular dependency has been identified in the Dependency Injection (DI) system of your Angular application.
-
-## Underlying Causes of NG0200 Error
-
-Circular dependencies occur when two or more classes or modules depend on each other, creating a loop in their dependencies. The most common causes of Angular Error NG0200 include:
-
-1. **Class Interdependence:** When two or more classes have direct dependencies on each other, forming a circular chain.
-
-2. **Module Relationships:** Circular dependencies can also arise at the module level, where modules reference each other in a loop.
-
-## Implications of Circular Dependencies
-
-Circular dependencies can lead to several adverse effects within your Angular application:
-
-1. **Reduced Readability:** Code with circular dependencies becomes harder to comprehend, increasing the likelihood of errors.
-
-2. **Maintenance Challenges:** Debugging and maintaining code with circular dependencies can be time-consuming and error-prone.
-
-3. **Potential Bugs:** Circular dependencies might introduce subtle bugs that are difficult to identify and rectify.
-
-## Strategies to Resolve NG0200 Error
-
-Overcoming Angular Error NG0200 requires a strategic approach to break the circular dependency chain:
-
-1. **Dependency Inversion Principle:** Adhere to the Dependency Inversion Principle, which promotes the use of interfaces or abstractions to decouple components and avoid direct class dependencies.
-
-2. **Refactor Modules:** Consider restructuring your application's modules to eliminate circular references between them.
-
-3. **Mediator Design Pattern:** Introduce a mediator or facade design pattern to mediate communication between components without direct dependencies.
-
-## Example Resolution
-
-Suppose you encounter the NG0200 error due to circular dependencies between `ServiceA` and `ServiceB`. Here's how you might resolve it:
-
-```typescript
-// Before
-// ServiceA.ts
-import { ServiceB } from "./serviceB";
-
-// ServiceB.ts
-import { ServiceA } from "./serviceA";
-
-// After (Applying Dependency Inversion Principle)
-// IService.ts
-interface IService {}
-
-// ServiceA.ts
-import { IService } from "./iService";
-import { ServiceB } from "./serviceB";
-
-// ServiceB.ts
-import { IService } from "./iService";
-import { ServiceA } from "./serviceA";
-```
-
-## Conclusion
-
-Angular Error NG0200, indicating the presence of circular dependencies, might pose a challenge, but with a clear understanding of its origins and consequences, you can effectively address it. By applying principles like Dependency Inversion, modular restructuring, and design patterns, you'll be on the path to a more maintainable and robust Angular application.
-
-Remember, reducing the complexity of your application's dependencies not only resolves errors but also enhances the overall quality of your codebase.
-
-Happy coding! 🛠️
-
-**Note:** The information provided here is based on knowledge available up to September 2021. For the most up-to-date information, refer to the official Angular documentation and relevant community resources.
