--- src/content/errors/0302.md
@@ -1,86 +1,0 @@
----
-title: "0302"
-slug: "0302"
-description: Angular Error NG0302 Multiple Components Matched
-author: Rafael Mestre
-authorTwitter: mestre_dev
-lastUpdated: Apr 12, 2021
----
-
-# Navigating Angular Error NG0302: Multiple Components Matched
-
-As an Angular developer, you're no stranger to the intricacies of building dynamic web applications. However, along the development journey, you might encounter various challenges, including cryptic error messages. In this article, we'll delve into Angular Error NG0302, which relates to multiple components matching a given route, and we'll explore its meaning, causes, and strategies for effective resolution.
-
-## Demystifying NG0302 Error
-
-Angular Error NG0302 is a common routing error that arises when the Angular router encounters multiple components that could potentially match a single route. The error message might appear like this:
-
-```
-Error NG0302: Found multiple component matches for route from '/example' to 'ComponentA, ComponentB'.
-```
-
-While this error message might seem perplexing, let's break it down to grasp its essence.
-
-## Understanding the Error Message
-
-1. **`Error NG0302:`** This unique identifier helps developers quickly recognize the specific error at hand.
-
-2. **`Found multiple component matches for route from '/example' to 'ComponentA, ComponentB'.`** This part of the error message indicates that the Angular router detected multiple components that could potentially match the specified route, causing ambiguity.
-
-## Root Causes of NG0302 Error
-
-Angular Error NG0302 occurs due to a clash in component routing, usually stemming from one of these scenarios:
-
-1. **Misconfigured Routes:** When you have multiple routes defined in your Angular application that unintentionally lead to the same or similar component.
-
-2. **Route Parameter Overlaps:** If you have routes with parameters that overlap, the router might interpret them as similar routes.
-
-3. **Wild Card Routes:** Usage of wild card (`**`) routes inappropriately can lead to conflicts.
-
-## Implications of Multiple Component Matches
-
-Encountering NG0302 errors can lead to several challenges within your Angular application:
-
-1. **Routing Confusion:** Ambiguous routes can cause unexpected navigation behavior, leading users to incorrect or unintended pages.
-
-2. **Debugging Complexity:** Resolving routing issues, especially when involving multiple components, can be time-consuming and challenging.
-
-3. **User Experience Impact:** Inconsistent routing can negatively affect the user experience, leading to confusion and frustration.
-
-## Strategies for NG0302 Error Resolution
-
-Addressing Angular Error NG0302 requires a methodical approach to eliminate the ambiguity in your route configurations:
-
-1. **Review Route Configurations:** Thoroughly inspect your route definitions to identify redundant or overlapping routes.
-
-2. **Parameter Naming:** Ensure that route parameters have distinct names to prevent clashes between different routes.
-
-3. **Specificity:** Aim for route definitions that are as specific as possible to avoid inadvertently matching unintended components.
-
-## Example Resolution
-
-Suppose you encounter NG0302 due to multiple routes pointing to `ComponentA` and `ComponentB`. Here's how you might resolve it:
-
-```typescript
-// Before (Misconfigured Routes)
-const routes: Routes = [
-  { path: "example", component: ComponentA },
-  { path: "example", component: ComponentB },
-];
-
-// After (Distinct Routes)
-const routes: Routes = [
-  { path: "routeA", component: ComponentA },
-  { path: "routeB", component: ComponentB },
-];
-```
-
-## Conclusion
-
-Angular Error NG0302, indicating multiple component matches for a route, might seem daunting, but with a clear understanding of its causes and implications, you can effectively tackle it. By reviewing route configurations, ensuring parameter distinctiveness, and focusing on route specificity, you'll pave the way for a more seamless and error-free routing experience.
-
-Remember, clarity in route definitions not only resolves errors but also contributes to a smoother navigation flow and enhanced user satisfaction.
-
-Happy coding! 🌐
-
-**Note:** The information provided here is based on knowledge available up to September 2021. For the most up-to-date information, refer to the official Angular documentation and relevant community resources.
