--- src/content/errors/6002.md
@@ -1,74 +1,0 @@
----
-title: "6002"
-slug: "6002"
-description: Demystifying Angular Error NG6002 What You Need to Know
-author: Rafael Mestre
-authorTwitter: mestre_dev
-lastUpdated: Sep 12, 2021
----
-
-# Demystifying Angular Error NG6002: What You Need to Know
-
-If you're a developer working with Angular, you've likely encountered a variety of errors and warnings during your coding journey. One such error is NG6002, which might seem puzzling at first, but fear not! In this blog post, we'll dive into the depths of NG6002 and unravel its meaning, causes, and potential solutions.
-
-## Understanding NG6002 Error
-
-The NG6002 error is a specific type of error that occurs within Angular applications. The error message itself might look something like this:
-
-```
-Error NG6002: Appears in the NgModule.imports of AppModule, but could not be resolved to an NgModule class.
-```
-
-At first glance, this error message might seem a bit cryptic, but let's break it down into more understandable terms.
-
-## Decoding the Error Message
-
-1. **`Error NG6002:`** This is the identifier for the error, which Angular uses to distinguish between different types of errors.
-
-2. **`Appears in the NgModule.imports of AppModule,`** This part of the error message tells you where the error was encountered. In this example, it's within the `imports` array of the `AppModule`.
-
-3. **`but could not be resolved to an NgModule class.`** This is the crucial piece of information. It indicates that Angular is trying to find a reference to a module (probably through an import), but it can't find a corresponding module class.
-
-## Common Causes of NG6002
-
-The NG6002 error typically occurs due to one or more of the following reasons:
-
-1. **Incorrect Import Path:** You might have mistyped the import path of a module in your code. This could be a typo in the module's file path or module name.
-
-2. **Missing Module Export:** If you're trying to import a module from a file that doesn't export any module, Angular won't be able to resolve it.
-
-3. **Circular Dependencies:** Circular dependencies occur when two or more modules depend on each other. This can confuse Angular's dependency resolution process and result in the NG6002 error.
-
-## Troubleshooting and Solutions
-
-1. **Check Import Paths:** Double-check all your import paths to ensure they are correct. Make sure the paths are relative to the file that's trying to import the module.
-
-2. **Verify Module Exports:** Ensure that the module you're trying to import actually exports a module class. This is usually done using the `@NgModule()` decorator.
-
-3. **Resolve Circular Dependencies:** If you suspect circular dependencies, refactor your code to break the circular chain. Consider creating a separate module for shared dependencies.
-
-## Example Fix
-
-Suppose you encounter the NG6002 error in your `AppModule` while trying to import a module named `SharedModule`. Here's how you might fix it:
-
-1. **Check Import Path:** Make sure the import path for `SharedModule` is correct, considering the relative path from your `AppModule` file.
-
-2. **Verify Module Exports:** Ensure that `SharedModule` has the `@NgModule()` decorator and that it's correctly exporting the module class.
-
-3. **Circular Dependencies:** If there are circular dependencies involving `SharedModule`, consider creating a new module to house the shared dependencies.
-
-```typescript
-// Incorrect import (may lead to NG6002)
-// import { SharedModule } from '../shared.module';
-
-// Correct import
-import { SharedModule } from "./path-to-shared.module";
-```
-
-## Wrapping Up
-
-Angular errors like NG6002 can be intimidating, but with a solid understanding of what they mean and some troubleshooting skills, you can confidently overcome them. Remember to meticulously review your import paths, verify module exports, and tackle circular dependencies if they're causing the issue. By following these steps, you'll be well on your way to a more robust and error-free Angular application.
-
-Happy coding! 😄🚀
-
-**Note:** The information provided here is based on knowledge available up to September 2021. For the most up-to-date information, refer to the official Angular documentation and relevant community resources.
