+++ vite.config.ts
@@ -15,15 +15,7 @@ export default defineConfig(({ mode }) => ({
   plugins: [
     analog({
       prerender: {
-        routes: [
-          '/',
-          '/errors',
-          '/errors/0100',
-          '/errors/0200',
-          '/errors/0302',
-          '/errors/6002',
-          '/errors-overview',
-        ],
+        routes: ['/', '/errors', '/errors/0100'],
       },
     }),
   ],
